+++ lib/propolis-client/src/instance_spec/openapi_impls.rs
